+++ packages/partial-markdown/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.4.0 — UNRELEASED
+## 0.4.0 — 2026-05-07
 
 ### Added